seeders y lifeinsirance

+++ backend/database/migrations/2014_15_12_000000_create_users_table.php
@@ -23,7 +23,7 @@ return new class extends Migration
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->float('creditLimit');
+            $table->decimal('creditLimit',15,2);
             $table->rememberToken();
             $table->timestamps();
         });
